assets/PDF Generator/requirements.txt
fastapi
uvicorn
sqlalchemy
pymupdf
python-multipart
jinja2
pandas
openpyxl
aiofiles
itsdangerous
qrcode
pillow
arabic-reshaper
python-bidi
